+++ src/modules/index.js
@@ -1,0 +1,4 @@
+export * as files from './files.js';
+export * as osModule from './ocCommands.js';
+export { hash } from './hash.js';
+export * as zipModule from './zipModule.js';
