+++ src/utils/utils.js
@@ -6,7 +6,9 @@ export const toCapitalizeText = (text) => {
 };
 export const greeting  = (username) => console.log(`Welcome to the File Manager, ${toCapitalizeText(username)} !`);
 export const goodbye  = (username) => console.log(`Thank you for using File Manager, ${toCapitalizeText(username)} goodbye!`);
-export const noElementMessage = () => console.log("no element in this folder plese try ls function");
+export const opearationFailedMessage = () => {
+  console.log('Operation failed');
+};
 export const tryCatchWrapper = async (fn, arg, argTwo=null) => {
       try {
         if (argTwo) {
@@ -14,13 +16,10 @@ export const tryCatchWrapper = async (fn, arg, argTwo=null) => {
         }
         return await fn(arg);
       } catch(e) {
-        console.log('Operation failed');
+        opearationFailedMessage();
     }
 };
 export const cretateNewFolder = (dest) => {
   !!Object.values(parse(dest))[1] &&
   mkdirSync(dirname(dest), { recursive: true });
-};
-export const opearationFailedMessage = () => {
-  console.log('Operation failed');
 };
